Refactor StorageImpl:
- PoolWrapper is passed as an external dependency
- Postgres initialization code is moved into PgConnectionInit class
- Corresponding fixes in Application and tests
