+++ test/system/irohad_test_data/config.sample.copy
@@ -1,0 +1,11 @@
+{
+    "block_store_path": "/var/folders/7m/bv1417y13dg5y4t0x04cc9dr0000gn/T/ebd6-ee80-3733-5c4c",
+    "torii_port": 50051,
+    "internal_port": 10001,
+    "pg_opt": "host=localhost port=5432 user=postgres password=mysecretpassword",
+    "max_proposal_size": 10,
+    "proposal_delay": 5000,
+    "vote_delay": 5000,
+    "mst_enable": false,
+    "mst_expiration_time": 1440
+}
